update : Water Tracking UI screen started (Not Finished)

+++ mobile/src/main/java/fit/asta/health/navigation/track/model/net/common/MultiGraphParent.kt
@@ -17,6 +17,6 @@ data class MultiGraphParent(
         @SerializedName("xVal")
         val xVal: List<String>,
         @SerializedName("yVal")
-        val yVal: List<Double>
+        val yVal: List<Float>
     )
 }
